Added a line to fix the bug with recycler view deleting.
Moved the deletePresentDialog into the presentsadapter class and tied it to the delete fab.
