+++ app/src/main/java/com/jj/honeypot/PresentsActivity.java
@@ -103,34 +103,4 @@ public class PresentsActivity extends AppCompatActivity {
         alertDialog.show();
     }
 
-    public void deletePresentDialog(final int presentIndex) {
-        // what happens when you click the delete fab next to
-        // a given present
-        AlertDialog.Builder dialog = new AlertDialog.Builder(this);
-        dialog.setTitle(R.string.dialog_delete_present_title);
-        dialog.setMessage(R.string.dialog_delete_present_message);
-        dialog.setPositiveButton(
-            R.string.dialog_delete_present_positive_button,
-            new DialogInterface.OnClickListener() {
-                @Override
-                public void onClick(DialogInterface dialog, int which) {
-                    boolean deleteSuccess = Presents.deletePresent(getAppFilesDir(), presentIndex);
-                    if (deleteSuccess) {
-                        finish();
-                        startActivity(getIntent());
-                    }
-                }
-            }
-        );
-        dialog.setNegativeButton(
-            R.string.dialog_delete_present_negative_button,
-            new DialogInterface.OnClickListener() {
-                @Override
-                public void onClick(DialogInterface dialog, int which) {
-                    return;
-                }
-            }
-        );
-    }
-
 }
