+++ app/src/main/java/com/jj/honeypot/PresentsAdapter.java
@@ -1,6 +1,7 @@
 package com.jj.honeypot;
 
 import android.content.Context;
+import android.content.DialogInterface;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -9,11 +10,15 @@ import android.widget.ImageView;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
+import androidx.appcompat.app.AlertDialog;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 
 import java.util.ArrayList;
+
+//import static java.security.AccessController.getContext;
+
 
 public class PresentsAdapter extends RecyclerView.Adapter<PresentsAdapter.ViewHolder> {
     public class ViewHolder extends RecyclerView.ViewHolder {
@@ -25,9 +30,9 @@ public class PresentsAdapter extends RecyclerView.Adapter<PresentsAdapter.ViewHo
         public FloatingActionButton deletePresent;
         public FloatingActionButton editPresent;
 
-        public ViewHolder(View view, String afd) {
+        public ViewHolder(View view, String appFilesDirPresentsAdapter) {
             super(view);
-            appFilesDir = afd;
+            appFilesDir = appFilesDirPresentsAdapter;
             name = (TextView) view.findViewById(R.id.textViewPresentName);
             price = (TextView) view.findViewById(R.id.textViewPresentPrice);
             store = (TextView) view.findViewById(R.id.textViewPresentStore);
@@ -40,9 +45,9 @@ public class PresentsAdapter extends RecyclerView.Adapter<PresentsAdapter.ViewHo
     private ArrayList<Present> presents;
     private String appFilesDir;
 
-    public PresentsAdapter(ArrayList<Present> inputPresents, String afd) {
+    public PresentsAdapter(ArrayList<Present> inputPresents, String appFilesDirAdapter) {
         presents = inputPresents;
-        appFilesDir = afd;
+        appFilesDir = appFilesDirAdapter;
     }
 
     @NonNull
@@ -55,8 +60,43 @@ public class PresentsAdapter extends RecyclerView.Adapter<PresentsAdapter.ViewHo
         return viewHolder;
     }
 
+    public void deletePresentDialog(final int presentIndex, Context context,final String appFilesDir) {
+        // what happens when you click the delete fab next to
+        // a given present
+        AlertDialog.Builder dialog = new AlertDialog.Builder(context);
+        dialog.setTitle(R.string.dialog_delete_present_title);
+        dialog.setMessage(R.string.dialog_delete_present_message);
+        dialog.setPositiveButton(
+                R.string.dialog_delete_present_positive_button,
+                new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialog, int which) {
+                        boolean deleteSuccess = Presents.deletePresent(appFilesDir, presentIndex);
+                        if (deleteSuccess) {
+                            presents.remove(presentIndex);
+                            notifyItemRemoved(presentIndex);
+                            //notifyItemRangeChanged is necessary to cover the instance when the
+                            //top item in the list is being removed.
+                            notifyItemRangeChanged(presentIndex,getItemCount());
+                        }
+                    }
+                }
+        );
+        dialog.setNegativeButton(
+                R.string.dialog_delete_present_negative_button,
+                new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialog, int which) {
+                        return;
+                    }
+                }
+        );
+        AlertDialog alertDialog = dialog.create();
+        alertDialog.show();
+    }
+
     @Override
-    public void onBindViewHolder(@NonNull PresentsAdapter.ViewHolder holder, final int position) {
+    public void onBindViewHolder(final @NonNull PresentsAdapter.ViewHolder holder, final int position) {
         Present present = presents.get(position);
         // avoid null present error when attempting
         // to get properties
@@ -78,16 +118,9 @@ public class PresentsAdapter extends RecyclerView.Adapter<PresentsAdapter.ViewHo
         holder.deletePresent.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Log.d("Justin","Starting delete");
-                Presents.deletePresent(appFilesDir,position);
-                Log.d("Justin","Removing at: "+position);
-                Log.d("Justin","presents.size(): "+presents.size());
-                presents.remove(position);
-                notifyItemRemoved(position);
-
+                deletePresentDialog(position,holder.name.getContext(),appFilesDir);
             }
         });
-//        holder.deletePresent.setOnTouchListener();
     }
 
     @Override
